Change Note ID and Label ID to UUID
The note id was already treated as a UUID, but stored as a string, which was not ideal. Now to label id has also been changed from an auto incrementing integer to a UUID.
The reason is that both of these entities can be created locally by the client and later uploaded to the remote database, which means they need an identifier for the local database. For now these locally generated ID's will also be used as keys in the remote database, but later this could be changed. Either way it's more convenient to treat both as UUID's.