Change Note model to only contain id's of labels
The plan is that label creation will always come first in the local->remote sync pipeline, so by the time the note gets updated with new labels, the labels corresponding to the label id's in the note object will already exist in the database.
The same works with remote->local synchronization (if the user adds new labels on another device for example): the client gets and updates its labels, so when the notes arrive, their label id's will already have their respective labels in the db.