Change refresh token management to handle unique devices
If the user/an attacker supplies an existing but invalidated refresh token, only those refresh tokens get deleted that are related to the device to which the invalidated refresh token was connected.

This means that in the case of an attack/user not using the app on one device the user will only get logged out of said device and not all of them.